+++ egui/src/containers/popup.rs
@@ -4,12 +4,22 @@ use crate::*;
 
 /// Show a tooltip at the current mouse position (if any).
 pub fn show_tooltip(ctx: &Arc<Context>, add_contents: impl FnOnce(&mut Ui)) {
-    if let Some(mouse_pos) = ctx.input().mouse.pos {
-        //  TODO: default size
-        let id = Id::tooltip();
-        let window_pos = mouse_pos + vec2(16.0, 16.0);
-        show_popup(ctx, id, window_pos, add_contents);
-    }
+    let tooltip_rect = ctx.memory().tooltip_rect;
+
+    let window_pos = if let Some(tooltip_rect) = tooltip_rect {
+        tooltip_rect.left_bottom()
+    } else if let Some(mouse_pos) = ctx.input().mouse.pos {
+        mouse_pos + vec2(16.0, 16.0)
+    } else {
+        return; // No good place for a tooltip :(
+    };
+
+    //  TODO: default size
+    let id = Id::tooltip();
+    let response = show_popup(ctx, id, window_pos, add_contents);
+
+    let tooltip_rect = tooltip_rect.unwrap_or_else(Rect::nothing);
+    ctx.memory().tooltip_rect = Some(tooltip_rect.union(response.rect));
 }
 
 /// Show a tooltip at the current mouse position (if any).