+++ egui/src/types.rs
@@ -116,7 +116,8 @@ impl std::fmt::Debug for Response {
 }
 
 impl Response {
-    /// Show this UI if the item was hovered (i.e. a tooltip)
+    /// Show this UI if the item was hovered (i.e. a tooltip).
+    /// If you call this multiple times the tooltips will stack underneath the previous ones.
     pub fn on_hover_ui(self, add_contents: impl FnOnce(&mut Ui)) -> Self {
         if self.hovered {
             crate::containers::show_tooltip(&self.ctx, add_contents);
@@ -124,7 +125,8 @@ impl Response {
         self
     }
 
-    /// Show this text if the item was hovered (i.e. a tooltip)
+    /// Show this text if the item was hovered (i.e. a tooltip).
+    /// If you call this multiple times the tooltips will stack underneath the previous ones.
     pub fn on_hover_text(self, text: impl Into<String>) -> Self {
         self.on_hover_ui(|ui| {
             ui.add(crate::widgets::Label::new(text));