+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Changed default font to [Ubuntu-Light](https://fonts.google.com/specimen/Ubuntu).
 * Remove minimum button width
 * Refactored `egui::Layout` substantially, changing its interface.
+* Calling ``on_hover_text`/`on_hover_ui` multiple times will stack tooltips underneath the previous ones.
 
 ### Removed 🔥